Install dev dependencies and link files
For this project, webpack and other dev dependencies are needed. This commit will install the dev dependencies required to allow us to run npx webpack and npx webpack serve, which are both integral to the development process. This commit also creates the webpack.config.js file, template.html file, src directory, index.js file, and links them all together so that we can get our project working properly. The next step is to work on the to do app itself