Make working tasks and projects
This commit will make tasks and projects that work. Meaning, when you submit a form for a new project, it is added to the display and when you click on the project, the tasks for that project will show. When you create a new task, it adds to the project you specify. Further work needs to be done to make tasks display dynamically update when you submit the new task form. Task and project elements also need functionality so you can edit and delete them.