Add date checking, exitable forms, deletable tasks, style to selectable items, display tasks.
This commit will make the task and project forms able to be exited. It will also check if the due dates of tasks are valid and alert/stop creation if invalid. The biggest features are the tasks which now display their details when clicked, and tasks that can now be deleted by clicking the x. The last things to do are to make projects deletable and make information save in local storage