Make project items deletable and validate projects
This commit will make it so that projects are removable both from the display and the projects array. When projects are removed, the display updates as well. This commit also validates submissions of the project form, checking if the requested project already exists. This prevents multiple projects of the same name being created and all having the same tasks.